Update plotting scripts
