+++ scripts/fig5.py
@@ -8,52 +8,242 @@
 # https://opensource.org/licenses/MIT.
 """This module defines functions used to generate images for figure 5.
 
-Note in our final paper, these results appear in the "Figure 6", after "Rapid Inference
-with Model-Free and Model-Based Policies".
+Panel A: Curvature Guided Policy
+ - `plot_curvature_guided_policy()`
 
-Panel B: 8-patch view finder
- - `plot_8lm_patches()`
+Panel B: Performance
+ - `plot_performance()`
 
-Panel C: Steps
- - `plot_steps()`
+Panel C and E: Hypotheses Visualizations
+ - `plot_object_hypotheses_visualization()`
+ - `plot_pose_hypotheses_visualization()`
 
-Panel D: Accuracy
- - `plot_accuracy()`
+Panels D and F: Hypotheses Evidence
+ - `plot_object_hypotheses_evidence()`
+ - `plot_pose_hypotheses_evidence()`
 
 Running the above functions requires that the following experiments have been run:
- - `fig5_visualize_8lm_patches`
- - `dist_agent_1lm_randrot_noise`
- - `dist_agent_2lm_randrot_noise`
- - `dist_agent_4lm_randrot_noise`
- - `dist_agent_8lm_randrot_noise`
- - `dist_agent_16lm_randrot_noise`
+ - `fig5_curvature_guided_policy`
+ - `fig5_hypothesis_driven_policy`
+ - `dist_agent_1lm_randrot_noise_nohyp
+ - `surf_agent_1lm_randrot_noise_nohyp`
+ - `surf_agent_1lm_randrot_noise`
 """
 
-from typing import Iterable
+from numbers import Number
+from typing import (
+    Iterable,
+    List,
+    Mapping,
+    Optional,
+    Tuple,
+)
 
 import matplotlib.pyplot as plt
 import numpy as np
-import pandas as pd
-import skimage
 from data_utils import (
     DMC_ANALYSIS_DIR,
     VISUALIZATION_RESULTS_DIR,
     DetailedJSONStatsInterface,
-    get_frequency,
-    load_eval_stats,
+    ObjectModel,
+    aggregate_1lm_performance_data,
+    load_object_model,
 )
+from matplotlib.lines import Line2D
+from matplotlib.patches import Polygon
 from plot_utils import (
     TBP_COLORS,
+    SensorModuleData,
+    axes3d_clean,
     axes3d_set_aspect_equal,
+    extract_style,
     init_matplotlib_style,
+    update_style,
     violinplot,
 )
+from scipy.spatial.transform import Rotation as R
 
 init_matplotlib_style()
-np.random.seed(0)
 
 OUT_DIR = DMC_ANALYSIS_DIR / "fig5"
 OUT_DIR.mkdir(parents=True, exist_ok=True)
+
+"""
+Style utilities. Can be hard-coded later when we decide on colors, etc.
+"""
+
+
+# Plotting styles.
+HYPOTHESIS_COLORS = [
+    TBP_COLORS["purple"],
+    TBP_COLORS["green"],
+    TBP_COLORS["blue"],
+    TBP_COLORS["pink"],
+    TBP_COLORS["yellow"],
+]
+
+STYLE = {
+    "target.color": "gray",
+    "target.alpha": 0.5,
+    "target.marker": "o",
+    "target.s": 1,
+    "target.edgecolor": "none",
+    "top_mlh.color": HYPOTHESIS_COLORS[0],
+    "top_mlh.alpha": 0.20,
+    "top_mlh.marker": "o",
+    "top_mlh.s": 1,
+    "top_mlh.edgecolor": "none",
+    "second_mlh.color": HYPOTHESIS_COLORS[1],
+    "second_mlh.alpha": 0.20,
+    "second_mlh.marker": "o",
+    "second_mlh.s": 1,
+    "second_mlh.edgecolor": "none",
+    "goal.color": TBP_COLORS["yellow"],
+    "goal.alpha": 1,
+    "goal.marker": "v",
+    "goal.s": 20,
+    "goal.edgecolor": "black",
+    "goal.lw": 0.5,
+    "goal.zorder": 20,
+}
+
+"""
+--------------------------------------------------------------------------------
+Panel A: Curvature Guided Policy
+--------------------------------------------------------------------------------
+"""
+
+
+def plot_curvature_guided_policy():
+    """Plot the object being sensed and the sensor path, including observations.
+
+    Requires the experiment `fig5_curvature_guided_policy` to have been run.
+
+    Output is saved to `DMC_ANALYSIS_DIR/fig5/curvature_guided_policy`.
+
+    """
+    # Initialize output directory.
+    out_dir = OUT_DIR / "curvature_guided_policy"
+    out_dir.mkdir(parents=True, exist_ok=True)
+
+    # Load the stats.
+    exp_dir = VISUALIZATION_RESULTS_DIR / "fig5_curvature_guided_policy"
+    detailed_stats_path = exp_dir / "detailed_run_stats.json"
+    detailed_stats_interface = DetailedJSONStatsInterface(detailed_stats_path)
+    stats = detailed_stats_interface[0]
+
+    # Initialize figure.
+    fig, ax = plt.subplots(1, 1, figsize=(3, 3), subplot_kw={"projection": "3d"})
+
+    # Plot the ground truth object.
+    model = load_object_model("dist_agent_1lm", "mug")
+    ax.scatter(
+        model.x,
+        model.y,
+        model.z,
+        color=model.rgba,
+        alpha=0.35,
+        s=4,
+        edgecolor="none",
+    )
+
+    # Plot the patches.
+    sm = SensorModuleData(stats["SM_0"])
+    for step in range(14):
+        sm.plot_raw_observation(ax, step, scatter=True, contour=True)
+
+    # Plot the sensor path.
+    sm.plot_sensor_path(ax, steps=14, start=False, scatter=False)
+
+    # Set axes properties.
+    ax.set_proj_type("persp", focal_length=0.5)
+    axes3d_clean(ax)
+    axes3d_set_aspect_equal(ax)
+    ax.view_init(elev=54, azim=-36, roll=60)
+
+    # Save the figure.
+    fig.savefig(out_dir / "curvature_guided_policy.png", bbox_inches="tight")
+    fig.savefig(out_dir / "curvature_guided_policy.svg", bbox_inches="tight")
+    plt.show()
+
+
+"""
+--------------------------------------------------------------------------------
+Panel B: Performance
+--------------------------------------------------------------------------------
+"""
+
+
+def plot_performance():
+    """Plot performance as a function of motor policy.
+
+    Requires the experiments `dist_agent_1lm_randrot_noise_nohyp`,
+    `surf_agent_1lm_randrot_noise_nohyp`, and `surf_agent_1lm_randrot_noise` to
+    have been run.
+
+    Output is saved to `DMC_ANALYSIS_DIR/fig5/performance`.
+
+    """
+    # Initialize output directory.
+    out_dir = OUT_DIR / "performance"
+    out_dir.mkdir(parents=True, exist_ok=True)
+
+    experiments = [
+        "dist_agent_1lm_randrot_noise_nohyp",
+        "surf_agent_1lm_randrot_noise_nohyp",
+        "surf_agent_1lm_randrot_noise",
+    ]
+    performance = aggregate_1lm_performance_data(experiments)
+
+    fig, axes = plt.subplots(1, 2, figsize=(3.65, 2.55))
+    xticks = np.arange(len(experiments))
+    xticklabels = [
+        "Random Walk",
+        "Model-Free",
+        "Model-Based",
+    ]
+
+    # Plot accuracy.
+    ax = axes[0]
+    ax.bar(xticks, performance["percent.correct"], width=0.8, color=TBP_COLORS["blue"])
+    ax.bar(
+        xticks,
+        performance["percent.correct_mlh"],
+        bottom=performance["percent.correct"],
+        width=0.8,
+        color=TBP_COLORS["blue"],
+        alpha=0.3,
+    )
+    ax.set_xticks(xticks)
+    ax.set_xticklabels(xticklabels, rotation=45, ha="center")
+    ax.set_ylabel("Accuracy (%)")
+    ax.set_ylim(0, 100)
+    ax.legend(["Converged", "Timed-Out"], loc="lower right", framealpha=1)
+
+    # Plot number of steps.
+    ax = axes[1]
+    violinplot(
+        performance["n_steps"],
+        xticks,
+        color=TBP_COLORS["blue"],
+        showmedians=True,
+        median_style=dict(color="lightgray", lw=2),
+        bw_method=0.1,
+        ax=ax,
+    )
+
+    ax.set_xticks(xticks)
+    ax.set_xticklabels(xticklabels, rotation=45, ha="center")
+    ax.set_ylabel("Number of Steps")
+    ax.set_ylim(0, 500)
+    fig.tight_layout()
+    fig.savefig(out_dir / "performance.png", bbox_inches="tight")
+    fig.savefig(out_dir / "performance.svg", bbox_inches="tight")
+    plt.show()
+
+    # Save the performance table.
+    performance = performance.drop(columns=["n_steps", "rotation_error"])
+    performance.to_csv(out_dir / "performance.csv")
 
 
 """
@@ -63,580 +253,825 @@ Utilities
 """
 
 
-def reduce_eval_stats(eval_stats: pd.DataFrame) -> pd.DataFrame:
-    """Reduce the eval stats dataframe to a single row per episode.
-
-    The main purpose of this function is to classify an episode as either "correct"
-    or "confused" based on the number of correct and confused LMs in an episode (or
-    "correct_mlh" and "confused_mlh" for timed-out episodes).
+def get_goal_states(stats: Mapping, achieved: Optional[bool] = True) -> List[Mapping]:
+    """Get the goal states for an episode.
 
     Args:
-        eval_stats: The eval stats dataframe.
+        stats: Detailed stats for an episode.
+        achieved: Whether to filter goal states by whether they were achieved. Default
+          is True.
 
     Returns:
-        pd.DataFrame: A dataframe with a single row per episode.
+        List[Mapping]: The goal states for an episode.
     """
-    performance_options = (
-        "correct",
-        "confused",
-        "no_match",
-        "correct_mlh",
-        "confused_mlh",
-        "time_out",
-        "pose_time_out",
-        "no_label",
-        "patch_off_object",
-    )
-
-    episodes = np.arange(eval_stats.episode.max() + 1)
-    assert np.array_equal(eval_stats.episode.unique(), episodes)  # sanity check
-    n_episodes = len(episodes)
-
-    # Columns of output dataframe. More are added later.
-    output_data = {
-        "primary_performance": np.zeros(n_episodes, dtype=object),
-    }
-    for name in performance_options:
-        output_data[f"n_{name}"] = np.zeros(n_episodes, dtype=int)
-
-    episode_groups = eval_stats.groupby("episode")
-    for episode, df in episode_groups:
-        # Find one result given many LM results.
-        row = {}
-
-        perf_counts = {key: 0 for key in performance_options}
-        perf_counts.update(df.primary_performance.value_counts())
-        # found = []
-        for name in performance_options:
-            row[f"n_{name}"] = perf_counts[name]
-
-        # Decide performance based on the number of correct/confused LM and
-        # correct_mlh/confused_mlh LMs. The rules are as follows:
-        #
-        # 1. If there are any correct or confused LMs (i.e., the episode did not time-out),
-        #    then we apply the following rules.
-        #   1.1 |correct LMs| > |confused-LMs| : correct
-        #   2.1 |correct LMs| < |confused-LMs| : confused
-        #   3.1 |correct LMs| = |confused-LMs| : we break the tie by comparing LMs
-        #      that timed out.
-        #      3.1.1 |correct_mlh LMs| > |confused_mlh LMs| : correct
-        #      3.2.1 |correct_mlh LMs| < |confused_mlh LMs| : confused
-        #      3.3.1 |correct_mlh LMs| = |confused_mlh LMs| : we break the tie by
-        #           randomly choosing correct or confused.
-        #
-        # 2. If the episode timed out, then we apply the following rules:
-        #   2.1. |correct_mlh LMs| > |confused_mlh LMs| : correct
-        #   2.2. |correct_mlh LMs| < |confused_mlh LMs| : confused
-        #   2.3. |correct_mlh LMs| = |confused_mlh LMs| : we break the tie by
-        #        randomly choosing correct_mlh or confused_mlh.
-        #
-        # 3. In the highly unusual case where no LMs terminated with
-        #    correct[_mlh] or confused[_mlh], then we assign performance based on
-        #    precedence rules of the remaining options. The following is a list
-        #    of remaining options in order of precedence (highest to lowest):
-        #     - time_out
-        #     - pose_time_out
-        #     - no_label
-        #     - patch_off_object
-        #
-
-        # Case 1: There are some correct or confused LMs.
-        if row["n_correct"] > 0 or row["n_confused"] > 0:
-            if row["n_correct"] > row["n_confused"]:
-                performance = "correct"
-            elif row["n_correct"] < row["n_confused"]:
-                performance = "confused"
-            else:
-                # Break the tie by comparing timed-out LMs.
-                if row["n_correct_mlh"] > row["n_confused_mlh"]:
-                    performance = "correct"
-                elif row["n_correct_mlh"] < row["n_confused_mlh"]:
-                    performance = "confused"
-                else:
-                    # Break the tie by flipping a coin.
-                    if np.random.rand() < 0.5:
-                        performance = "correct"
-                    else:
-                        performance = "confused"
-
-        # Case 2: The episode timed out.
-        elif row["n_correct_mlh"] > 0 or row["n_confused_mlh"] > 0:
-            if row["n_correct_mlh"] > row["n_confused_mlh"]:
-                performance = "correct_mlh"
-            elif row["n_correct_mlh"] < row["n_confused_mlh"]:
-                performance = "confused_mlh"
-            else:
-                # Break the tie by flipping a coin.
-                if np.random.rand() < 0.5:
-                    performance = "correct_mlh"
-                else:
-                    performance = "confused_mlh"
-
-        # Case 3: No LMs terminated with correct[_mlh] or confused[_mlh].
-        else:
-            # Apply precedence rules.
-            performance = None
-            for name in ("time_out", "pose_time_out", "no_label", "patch_off_object"):
-                if row[f"n_{name}"] > 0:
-                    performance = name
-                    break
-            if performance is None:
-                raise ValueError("No valid performance value found")
-
-        row["primary_performance"] = performance
-
-        for key, val in row.items():
-            output_data[key][episode] = val
-
-    # Add episode data not specific to the LM.
-    output_data["monty_matching_steps"] = (
-        episode_groups.monty_matching_steps.first().values
-    )
-    output_data["primary_target_object"] = (
-        episode_groups.primary_target_object.first().values
-    )
-    output_data["primary_target_rotation"] = (
-        episode_groups.primary_target_object.first().values
-    )
-    output_data["episode"] = episode_groups.episode.first().values
-    output_data["epoch"] = episode_groups.epoch.first().values
-
-    out = pd.DataFrame(output_data)
-    return out
+    goal_states = stats["LM_0"]["goal_states"]
+    if achieved is not None:
+        goal_states = [gs for gs in goal_states if gs["info"]["achieved"] == achieved]
+    possible_matches = stats["LM_0"]["possible_matches"]
+    for gs in goal_states:
+        step = gs["info"]["matching_step_when_output_goal_set"]
+        gs["info"]["is_pose_hypothesis"] = len(possible_matches[step]) == 1
+        gs["info"]["possible_matches"] = possible_matches[step]
+    return goal_states
 
 
-def aggregate_multilm_performance_data(experiments: Iterable[str]) -> pd.DataFrame:
-    """Save the performance table for the single LM experiments.
+def draw_triangle(
+    ax: plt.Axes,
+    x: Number,
+    y: Number,
+    theta: Number = 0,
+    degrees: bool = True,
+    radius: Number = 1,
+    align: str = "center",
+    **style,
+) -> Polygon:
+    """Draw a triangle, properly sized and located.
 
-    Output is saved to `DMC_ANALYSIS_DIR/fig3/performance/single_lm_performance.csv`.
+    Matplotlib wasn't playing nicely when trying to put triangles on the figure
+    that went outside the axes limits. This function converts the triangle to
+    normalized axes coordinates, so that it can be drawn on any axes.
+
+    Args:
+        ax: Axes to plot in.
+        x: x position in data coordinates.
+        y: y position in data coordinates.
+        theta: Counter-clockwise rotation of triangle. Default is 0, which has the
+            triangle pointing up.
+        degrees: Whether theta is in degrees. Defaults to True.
+        radius: Radius of the circle that inscribes the triangle in millimeters.
+        align: Alignment of triangle. Defaults to "center".
+        style: Additional style arguments for the matplotlib polygon.
+    Raises:
+        ValueError: If align is not one of "center", "bottom", "top", "left", or
+        "right".
+
+    Returns:
+        np.ndarray: Triangle vertices in normalized axes coordinates.
     """
+    t = np.arange(0, 1, 1 / 3)
+    x_coords = radius * np.cos(t * 2 * np.pi)
+    y_coords = radius * np.sin(t * 2 * np.pi)
+    verts_mm = np.stack([x_coords, y_coords], axis=1)
 
-    columns = {
-        "accuracy": [],
-        "percent.correct": [],
-        "percent.correct_mlh": [],
-        "n_steps": [],
-        "n_steps.mean": [],
-        "n_steps.median": [],
-        "rotation_error": [],
-        "rotation_error.mean": [],
-        "rotation_error.median": [],
+    # Get triangle pointing up.
+    d_theta = 2 * np.pi / 3 - np.pi / 2
+    c, s = np.cos(d_theta), np.sin(d_theta)
+    rot = np.array([[c, -s], [s, c]])
+    verts_mm = verts_mm @ rot
+
+    # Apply counter-clockwise rotation.
+    if degrees:
+        theta = np.deg2rad(-theta)
+    c, s = np.cos(theta), np.sin(theta)
+    rot = np.array([[c, -s], [s, c]])
+    verts_mm = verts_mm @ rot
+
+    # Get points in axis-length coordinates.
+    verts_pix = verts_mm * ax.figure.dpi / 25.4
+    bbox = ax.get_window_extent()
+    pix_per_ax_length = np.array([bbox.width, bbox.height])
+    verts_ax = verts_pix / pix_per_ax_length[None, :]
+
+    # Align triangle.
+    if align == "bottom":
+        offset = np.array([0, verts_ax[:, 1].min()])
+    elif align == "top":
+        offset = np.array([0, verts_ax[:, 1].max()])
+    elif align == "left":
+        offset = np.array([verts_ax[:, 0].min(), 0])
+    elif align == "right":
+        offset = np.array([verts_ax[:, 0].max(), 0])
+    elif align == "center":
+        offset = np.array([0, 0])
+    else:
+        raise ValueError(f"Invalid align: {align}")
+    verts_ax = verts_ax - offset[None, :]
+
+    # Move vertices to location on axes.
+    x_lim = ax.get_xlim()
+    y_lim = ax.get_ylim()
+    x_ax = (x - x_lim[0]) / (x_lim[1] - x_lim[0])
+    y_ax = (y - y_lim[0]) / (y_lim[1] - y_lim[0])
+    offset = np.array([x_ax, y_ax])
+    verts_ax = verts_ax + offset[None, :]
+
+    kwargs = {
+        "closed": True,
+        "transform": ax.transAxes,
+        "clip_on": False,
     }
-    for exp in experiments:
-        eval_stats = load_eval_stats(exp)
-        reduced_stats = reduce_eval_stats(eval_stats)
-        accuracy = 100 * get_frequency(
-            reduced_stats["primary_performance"], ("correct", "correct_mlh")
-        )
-        percent_correct = 100 * get_frequency(
-            reduced_stats["primary_performance"], "correct"
-        )
-        percent_correct_mlh = 100 * get_frequency(
-            reduced_stats["primary_performance"], "correct_mlh"
-        )
-        n_steps = eval_stats["num_steps"]
-        rotation_error = np.degrees(eval_stats.rotation_error.dropna())
+    kwargs.update(style)
+    triangle = Polygon(verts_ax, **kwargs)
+    ax.add_patch(triangle)
 
-        columns["accuracy"].append(accuracy)
-        columns["percent.correct"].append(percent_correct)
-        columns["percent.correct_mlh"].append(percent_correct_mlh)
-
-        columns["n_steps"].append(n_steps)
-        columns["n_steps.mean"].append(n_steps.mean())
-        columns["n_steps.median"].append(n_steps.median())
-
-        columns["rotation_error"].append(rotation_error)
-        columns["rotation_error.mean"].append(rotation_error.mean())
-        columns["rotation_error.median"].append(rotation_error.median())
-
-    return pd.DataFrame(columns, index=experiments)
+    return triangle
 
 
 """
 --------------------------------------------------------------------------------
-Panel B: 8-patch view finder
+Panels C and E: Hypothesis Visualizations
 --------------------------------------------------------------------------------
 """
 
 
-def plot_8lm_patches():
-    """Plot the 8-SM + view_finder visualization for figure 4.
+def get_mlh_for_object(object_name: str, stats: Mapping, step: int) -> Mapping:
+    """Get the most likely hypothesis for a given object.
 
-    Uses data from the experiment `fig4_visualize_8lm_patches` defined in
-    `configs/visualizations.py`. This function renders the sensor module
-    RGBA data in the scene (in 3D) and overlays the sensor module's patch
-    boundaries.
+    Args:
+        object_name (str): The object name.
+        stats (Mapping): Detailed stats for an episode.
+        step (int): The step to get the MLH for.
 
-    Output is saved to `DMC_ANALYSIS_DIR/fig5/8lm_patches`.
+    Returns:
+        Mapping: The MLH.
+    """
+    evidences = stats["LM_0"]["evidences"][step]
+    locations = stats["LM_0"]["possible_locations"][step]
+    rotations = stats["LM_0"]["possible_rotations"][0]
+    mlh_id = np.argmax(evidences[object_name])
+    return {
+        "object_name": object_name,
+        "mlh_id": mlh_id,
+        "evidence": evidences[object_name][mlh_id],
+        "location": np.array(locations[object_name][mlh_id]),
+        "rotation": R.from_matrix(rotations[object_name][mlh_id]),
+    }
+
+
+def get_top_two_mlhs(stats, step) -> Tuple[Mapping, Mapping]:
+    """Get the top two MLHs for a given step (different objects).
+
+    Args:
+        stats (Mapping): Detailed stats for an episode.
+        step (int): The step to get the MLHs for.
+
+    Returns:
+        Tuple[Mapping, Mapping]: The top two MLHs in descending order of evidence.
+    """
+    evidences = stats["LM_0"]["evidences"][step]
+    mlh_info = []
+    for object_name in evidences.keys():
+        mlh_info.append(get_mlh_for_object(object_name, stats, step))
+    lst = sorted(mlh_info, key=lambda x: x["evidence"], reverse=True)
+    return lst[0], lst[1]
+
+
+def get_top_two_mlhs_for_object(
+    object_name: str, stats: Mapping, step: int
+) -> Tuple[Mapping, Mapping]:
+    """Get the top two pose hypotheses for a given object.
+
+    Args:
+        object_name (str): The object/graph id.
+        stats (Mapping): Detailed stats for an episode.
+        step (int): The step to get the MLHs for.
+
+    Returns:
+        Tuple[Mapping, Mapping]: An object's top two pose hypothesis MLHs in
+        descending order of evidence.
+    """
+    evidences = stats["LM_0"]["evidences"][step]
+    locations = stats["LM_0"]["possible_locations"][step]
+    rotations = stats["LM_0"]["possible_rotations"][0]
+    sort_order = np.argsort(evidences[object_name])[::-1]
+    mlhs = []
+    for mlh_id in sort_order[:2]:
+        mlhs.append(
+            {
+                "object_name": object_name,
+                "mlh_id": mlh_id,
+                "evidence": evidences[object_name][mlh_id],
+                "location": np.array(locations[object_name][mlh_id]),
+                "rotation": R.from_matrix(rotations[object_name][mlh_id]),
+            }
+        )
+    return mlhs
+
+
+def get_graph_for_hypothesis(
+    mlh: Mapping,
+    stats: Mapping,
+    step: int,
+    pretrained_model: str = "surf_agent_1lm",
+) -> ObjectModel:
+    """Get the graph of the MLH in the same reference frame as the sensed target object.
+
+    Args:
+        mlh: The MLH dictionary.
+        stats: Detailed stats for an episode.
+        step: The step to get the MLH for.
+        pretrained_model: The pretrained model to use. Default is "surf_agent_1lm".
+
+    Returns:
+        ObjectModel: The graph of the MLH in the same reference frame as
+          the sensed target object.
+    """
+    rotated_mlh_location = mlh["rotation"].inv().apply(mlh["location"])
+    sensor_location = np.array(
+        stats["SM_0"]["processed_observations"][step]["location"]
+    )
+    learned_graph = load_object_model(pretrained_model, mlh["object_name"])
+    graph = learned_graph.rotated(mlh["rotation"].inv())
+    graph -= rotated_mlh_location
+    graph += sensor_location
+    return graph
+
+
+def plot_hypotheses_visualization(
+    stats: Mapping,
+    step: int,
+    top_mlh: Mapping,
+    second_mlh: Mapping,
+    style: Optional[Mapping] = None,
+) -> Tuple[plt.Figure, plt.Axes]:
+    """Plot the hypotheses for a given step.
+
+    Args:
+        stats: Detailed stats for an episode.
+        step: The step to plot the hypotheses for.
+        top_mlh: The MLH with the highest evidence value.
+        second_mlh: The MLH with the second highest evidence value.
+        style: Optional style overrides for the plot items.
+
+    Returns:
+        Tuple[plt.Figure, plt.Axes]: The figure and axes.
+    """
+    style = update_style(STYLE, style)
+
+    # Get ground-truth object and its pose.
+    target_object = stats["target"]["primary_target_object"]
+    target_position = np.array(stats["target"]["primary_target_position"])
+    target_rotation = R.from_euler(
+        "xyz", stats["target"]["primary_target_rotation_euler"], degrees=True
+    )
+    learned_position = np.array([0, 1.5, 0])
+
+    # Load sensor locations.
+    sensor_locations = np.array(
+        [obs["location"] for obs in stats["SM_0"]["processed_observations"]]
+    )
+    sensor_locations = sensor_locations[: step + 1]
+
+    fig, axes = plt.subplots(1, 2, figsize=(5, 4), subplot_kw={"projection": "3d"})
 
     """
-    # Initialize output directory.
-    out_dir = OUT_DIR / "8lm_patches"
-    out_dir.mkdir(exist_ok=True, parents=True)
+    First plot has ground-truth object and sensor path.
+    """
+    ax = axes[0]
 
-    # Load the detailed stats.
-    experiment_dir = VISUALIZATION_RESULTS_DIR / "fig5_visualize_8lm_patches"
-    detailed_stats_path = experiment_dir / "detailed_run_stats.json"
-    detailed_stats_interface = DetailedJSONStatsInterface(detailed_stats_path)
-    stats = detailed_stats_interface[0]
-
-    def pull(sm_num: int):
-        """Helper function for extracting necessary sensor data."""
-        sm_dict = stats[f"SM_{sm_num}"]
-        # Extract RGBA sensor patch.
-        rgba_2d = np.array(sm_dict["raw_observations"][0]["rgba"]) / 255.0
-        n_rows, n_cols = rgba_2d.shape[0], rgba_2d.shape[1]
-
-        # Extract locations and on-object filter.
-        semantic_3d = np.array(sm_dict["raw_observations"][0]["semantic_3d"])
-        pos_1d = semantic_3d[:, 0:3]
-        pos_2d = pos_1d.reshape(n_rows, n_cols, 3)
-        on_object_1d = semantic_3d[:, 3].astype(int) > 0
-        on_object_2d = on_object_1d.reshape(n_rows, n_cols)
-
-        # Filter out points that aren't on-object. Yields a flat list of points/colors.
-        return rgba_2d, pos_2d, on_object_2d
-
-    # Create a 3D plot of the semantic point cloud
-    fig = plt.figure(figsize=(4, 4))
-    ax = fig.add_subplot(1, 1, 1, projection="3d")
-    ax.view_init(elev=90, azim=-90, roll=0)
-    ax.set_proj_type("persp", focal_length=0.125)
-    ax.dist = 4.55
-
-    # Render the view finder's RGBA data in the scene.
-    rgba_2d, pos_2d, on_object_2d = pull(8)
-    rows, cols = np.where(on_object_2d)
-    pos_valid_1d = pos_2d[on_object_2d]
-    rgba_valid_1d = rgba_2d[on_object_2d]
+    # Plot ground-truth object.
+    learned_graph = load_object_model("surf_agent_1lm", target_object)
+    target_graph = learned_graph - learned_position
+    target_graph = target_graph.rotated(target_rotation)
+    target_graph += target_position
+    target_style = extract_style(style, "target")
+    if target_style.get("color") == "rgba":
+        target_style["color"] = target_graph.rgba
     ax.scatter(
-        pos_valid_1d[:, 0],
-        pos_valid_1d[:, 1],
-        pos_valid_1d[:, 2],
-        c=rgba_valid_1d,
-        marker="o",
-        alpha=0.3,
-        zorder=5,
-        s=10,
-        edgecolors="none",
+        target_graph.x,
+        target_graph.y,
+        target_graph.z,
+        **target_style,
     )
 
-    # Render patches and patch boundaries for all sensors.
-    for i in range(8):
-        # Load sensor data.
-        rgba_2d, pos_2d, on_object_2d = pull(i)
-        rows, cols = np.where(on_object_2d)
-        pos_valid_1d = pos_2d[on_object_2d]
-        rgba_valid_1d = rgba_2d[on_object_2d]
+    # Plot sensor path on ground-truth object.
+    sm = SensorModuleData(stats["SM_0"], style=style)
+    sm.plot_sensor_path(ax, steps=step + 1)
 
-        # Render the patch.
-        ax.scatter(
-            pos_valid_1d[:, 0],
-            pos_valid_1d[:, 1],
-            pos_valid_1d[:, 2],
-            c=rgba_valid_1d,
-            marker="o",
-            alpha=1,
-            zorder=10,
-            edgecolors="none",
-            s=1,
-        )
+    """
+    Second plot has first and second MLHs.
+    """
+    ax = axes[1]
 
-        # Draw the patch boundaries (complicated).
-        n_rows, n_cols = on_object_2d.shape
-        row_mid, col_mid = n_rows // 2, n_cols // 2
-        n_pix_on_object = on_object_2d.sum()
+    # Plot first and second MLHs.
+    top_mlh["graph"] = get_graph_for_hypothesis(
+        top_mlh, stats, step, pretrained_model="surf_agent_1lm"
+    )
+    top_mlh_style = extract_style(style, "top_mlh")
+    if top_mlh_style.get("color") == "rgba":
+        top_mlh_style["color"] = top_mlh["graph"].rgba
+    ax.scatter(
+        top_mlh["graph"].x,
+        top_mlh["graph"].y,
+        top_mlh["graph"].z,
+        **top_mlh_style,
+    )
 
-        if n_pix_on_object == 0:
-            contours = []
-        elif n_pix_on_object == on_object_2d.size:
-            temp = np.zeros((n_rows, n_cols), dtype=bool)
-            temp[0, :] = True
-            temp[-1, :] = True
-            temp[:, 0] = True
-            temp[:, -1] = True
-            contours = [np.argwhere(temp)]
-        else:
-            contours = skimage.measure.find_contours(
-                on_object_2d, level=0.5, positive_orientation="low"
+    second_mlh["graph"] = get_graph_for_hypothesis(
+        second_mlh, stats, step, pretrained_model="surf_agent_1lm"
+    )
+    second_mlh_style = extract_style(style, "second_mlh")
+    if second_mlh_style.get("color") == "rgba":
+        second_mlh_style["color"] = second_mlh["graph"].rgba
+    ax.scatter(
+        second_mlh["graph"].x,
+        second_mlh["graph"].y,
+        second_mlh["graph"].z,
+        **second_mlh_style,
+    )
+
+    # Plot the goal state's target if possible.
+    goal_states = get_goal_states(stats)
+    gs = None
+    for g in goal_states:
+        if g["info"]["matching_step_when_output_goal_set"] == step:
+            gs = g
+    if gs:
+        proposed_surface_loc = gs["info"]["proposed_surface_loc"]
+        goal_style = extract_style(style, "goal")
+        for ax in axes:
+            ax.scatter(
+                proposed_surface_loc[0],
+                proposed_surface_loc[1],
+                proposed_surface_loc[2],
+                **goal_style,
             )
-            contours = [] if contours is None else contours
-
-        for ct in contours:
-            row_mid, col_mid = n_rows // 2, n_cols // 2
-
-            # Contour may be floating point (fractional indices from scipy). If so,
-            # round rows/columns towards the center of the patch.
-            if not np.issubdtype(ct.dtype, np.integer):
-                # Round towards the center.
-                rows, cols = ct[:, 0], ct[:, 1]
-                rows_new, cols_new = np.zeros_like(rows), np.zeros_like(cols)
-                rows_new[rows >= row_mid] = np.floor(rows[rows >= row_mid])
-                rows_new[rows < row_mid] = np.ceil(rows[rows < row_mid])
-                cols_new[cols >= col_mid] = np.floor(cols[cols >= col_mid])
-                cols_new[cols < col_mid] = np.ceil(cols[cols < col_mid])
-                ct_new = np.zeros_like(ct, dtype=int)
-                ct_new[:, 0] = rows_new.astype(int)
-                ct_new[:, 1] = cols_new.astype(int)
-                ct = ct_new
-
-            # Drop any points that happen to be off-object (it's possible that
-            # some boundary points got rounded off-object).
-            points_on_object = on_object_2d[ct[:, 0], ct[:, 1]]
-            ct = ct[points_on_object]
-
-            # In order to plot the boundary as a line, we need the points to
-            # be in order. We can order them by associating each point with its
-            # angle from the center of the patch. This isn't a general solution,
-            # but it works here.
-            Y, X = row_mid - ct[:, 0], ct[:, 1] - col_mid  # pixel to X/Y coords.
-            theta = np.arctan2(Y, X)
-            sort_order = np.argsort(theta)
-            ct = ct[sort_order]
-
-            # Finally, plot the contour.
-            xyz = pos_2d[ct[:, 0], ct[:, 1]]
-            ax.plot(xyz[:, 0], xyz[:, 1], xyz[:, 2], c="k", linewidth=3, zorder=20)
-
-    axes3d_set_aspect_equal(ax)
-    ax.axis("off")
-
-    fig.savefig(out_dir / "8lm_patches.png")
-    fig.savefig(out_dir / "8lm_patches.svg")
-    plt.show()
+    return fig, axes
 
 
-"""
---------------------------------------------------------------------------------
-Panel C: Steps
---------------------------------------------------------------------------------
-"""
+def get_visualization_legend_handles(
+    labels: List[str],
+    style: Mapping,
+) -> List[Line2D]:
+    style = update_style(STYLE, style)
+    legend_handles = []
+    for i, name in enumerate(("sensor_path.scatter", "top_mlh", "second_mlh", "goal")):
+        st = extract_style(style, name)
+        legend_handles.append(
+            Line2D(
+                [0],
+                [0],
+                marker=st["marker"],
+                color="w",
+                markerfacecolor=st["color"],
+                markeredgecolor=st["color"],
+                markersize=6,
+                label=labels[i],
+            )
+        )
+
+    return legend_handles
 
 
-def plot_steps():
-    """Plot the number of steps taken by 1-LM and multi-LM experiments.
+def plot_object_hypotheses_visualization():
+    """Draw top two object MLHs, sensor paths, and goal the state point.
 
-    Requires the following experiments to have been run:
-    - `dist_agent_1lm_randrot_noise`
-    - `dist_agent_2lm_randrot_noise`
-    - `dist_agent_4lm_randrot_noise`
-    - `dist_agent_8lm_randrot_noise`
-    - `dist_agent_16lm_randrot_noise`
+    Requires having run `fig5_hypothesis_driven_policy`.
 
-    Output is saved to `DMC_ANALYSIS_DIR/fig5/performance`.
-
+    Output is saved to `DMC_ANALYSIS_DIR/fig5/object_hypotheses`.
     """
-    # Initialize output directory.
-    out_dir = OUT_DIR / "performance"
-    out_dir.mkdir(exist_ok=True, parents=True)
+    out_dir = OUT_DIR / "object_hypotheses"
+    out_dir.mkdir(parents=True, exist_ok=True)
 
-    experiments = [
-        "dist_agent_1lm_randrot_noise",
-        "dist_agent_2lm_randrot_noise",
-        "dist_agent_4lm_randrot_noise",
-        "dist_agent_8lm_randrot_noise",
-        "dist_agent_16lm_randrot_noise",
-    ]
-    performance = aggregate_multilm_performance_data(experiments)
+    style = STYLE.copy()
+    experiment = "fig5_hypothesis_driven_policy"
+    episode = 0
 
-    fig = plt.figure(figsize=(3.4, 3))
-    gs = fig.add_gridspec(3, 1)  # 3 rows, bottom plot will take 2 rows
+    exp_dir = VISUALIZATION_RESULTS_DIR / experiment
+    detailed_stats_path = exp_dir / "detailed_run_stats.json"
+    detailed_stats_interface = DetailedJSONStatsInterface(detailed_stats_path)
+    stats = detailed_stats_interface[episode]
 
-    # Create the two subplots with shared x-axis
-    top_ax = fig.add_subplot(gs[0, 0])  # Top subplot takes 1/3
-    bottom_ax = fig.add_subplot(gs[1:, 0], sharex=top_ax)  # Bottom subplot takes 2/3
-    fig.subplots_adjust(hspace=0.05)
+    # Select a step where there's a goal state was achieved, and it was
+    # for distuinguishing between different objects.
+    goal_states = get_goal_states(stats, achieved=True)
+    goal_states = [g for g in goal_states if not g["info"]["is_pose_hypothesis"]]
+    gs = goal_states[0]
+    step = gs["info"]["matching_step_when_output_goal_set"]
 
-    # Plot params.
-    ylims = [(0, 110), (440, 500)]
-    yticks = [
-        [0, 25, 50, 75, 100],
-        [450, 475, 500],
-    ]
-    xticks = np.arange(5)
+    # Get the pose MLHs.
+    top_mlh, second_mlh = get_top_two_mlhs(stats, step)
 
-    # Plot distribution of n_steps for 1-LM
-    for ax_num, ax in enumerate([bottom_ax, top_ax]):
-        violinplot(
-            [performance.n_steps[0]],
-            [xticks[0]],
-            color=TBP_COLORS["blue"],
-            width=0.8,
-            showmedians=True,
-            median_style=dict(color="lightgray", lw=2),
-            bw_method=0.1,
-            ax=ax,
-        )
-
-    # Plot distribution of n_steps for >1 LM
-    for ax_num, ax in enumerate([bottom_ax, top_ax]):
-        violinplot(
-            performance.n_steps[1:],
-            xticks[1:],
-            color=TBP_COLORS["purple"],
-            width=0.8,
-            showmedians=True,
-            median_style=dict(color="lightgray", lw=2),
-            bw_method=0.1,
-            ax=ax,
-        )
-
-    # Plot a line connecting the means across all conditions.
-    means = performance["n_steps.mean"]
-    for ax in [bottom_ax, top_ax]:
-        ax.scatter(
-            xticks,
-            means,
-            color="black",
-            marker="o",
-            s=20,
-        )
-        ax.plot(xticks, means, color="black", linestyle="-", linewidth=2, zorder=10)
-
-    # Set y-limits and ticks.
-    for ax_num, ax in enumerate([bottom_ax, top_ax]):
-        ax.set_ylim(ylims[ax_num])
-        ax.set_yticks(yticks[ax_num])
-
-    # Sets parameters for both x-axes (they're shared, so removing ticks for the
-    # top plot removes ticks for the bottom plot).
-    bottom_ax.set_xlabel("Num. LMs")
-    bottom_ax.set_xticks(xticks)
-    bottom_ax.set_xticklabels(["1", "2", "4", "8", "16"])
-
-    bottom_ax.set_ylabel("Number of Steps")
-    top_ax.spines.bottom.set_visible(False)
-
-    # Draw y-axis divider markers.
-    marker_kwargs = dict(
-        marker=[(-1, -0.5), (1, 0.5)],
-        markersize=8,
-        linestyle="none",
-        color="k",
-        mec="k",
-        mew=1,
-        clip_on=False,
+    fig, axes = plot_hypotheses_visualization(
+        stats,
+        step,
+        top_mlh,
+        second_mlh,
+        style=style,
     )
-    top_ax.plot([0], [0], transform=top_ax.transAxes, **marker_kwargs)
-    bottom_ax.plot([0], [1], transform=bottom_ax.transAxes, **marker_kwargs)
 
-    fig.savefig(out_dir / "steps.png")
-    fig.savefig(out_dir / "steps.svg")
+    # Add label, legends, etc.
+    view_angles = [(-50, -180, 0), (-80, 180, 0)]
+    for i, ax in enumerate(axes):
+        ax.set_proj_type("persp", focal_length=0.8)
+        axes3d_set_aspect_equal(ax)
+        axes3d_clean(ax)
+        ax.view_init(*view_angles[i])
+
+    sm = SensorModuleData(stats["SM_0"], style=style)
+    legend_handles = get_visualization_legend_handles(
+        ["sensor", "spoon", "fork", "goal"], sm.style
+    )
+    axes[1].legend(
+        handles=legend_handles,
+        bbox_to_anchor=(0.1, 0.8),
+    )
+
+    fig.savefig(out_dir / "object_hypotheses_visualization.png", bbox_inches="tight")
+    fig.savefig(out_dir / "object_hypotheses_visualization.svg", bbox_inches="tight")
     plt.show()
 
-
-"""
---------------------------------------------------------------------------------
-Panel D: Accuracy
---------------------------------------------------------------------------------
-"""
+    return fig, axes
 
 
-def plot_accuracy():
-    """Plot accuracy of 1-LM and multi-LM experiments.
+def plot_pose_hypotheses_visualization():
+    """Draw top two pose MLHs, sensor paths, and goal the state point.
 
-    Requires the following experiments to have been run:
-    - `dist_agent_1lm_randrot_noise`
-    - `dist_agent_2lm_randrot_noise`
-    - `dist_agent_4lm_randrot_noise`
-    - `dist_agent_8lm_randrot_noise`
-    - `dist_agent_16lm_randrot_noise`
+    Requires having run `fig5_hypothesis_driven_policy`.
 
-    Output is saved to `DMC_ANALYSIS_DIR/fig5/performance`.
-
+    Output is saved to `DMC_ANALYSIS_DIR/fig5/pose_hypotheses`.
     """
-    # Initialize output directory.
-    out_dir = OUT_DIR / "performance"
-    out_dir.mkdir(exist_ok=True, parents=True)
+    out_dir = OUT_DIR / "pose_hypotheses"
+    out_dir.mkdir(parents=True, exist_ok=True)
 
-    experiments = [
-        "dist_agent_1lm_randrot_noise",
-        "dist_agent_2lm_randrot_noise",
-        "dist_agent_4lm_randrot_noise",
-        "dist_agent_8lm_randrot_noise",
-        "dist_agent_16lm_randrot_noise",
-    ]
-    performance = aggregate_multilm_performance_data(experiments)
+    style = update_style(STYLE, {"target.s": 4, "top_mlh.s": 4, "second_mlh.s": 4})
+    experiment = "fig5_hypothesis_driven_policy"
+    episode = 1
 
-    fig, axes = plt.subplots(2, 1, figsize=(3.4, 3), sharex=True)
-    top_ax, bottom_ax = axes
-    fig.subplots_adjust(hspace=0.05)
+    exp_dir = VISUALIZATION_RESULTS_DIR / experiment
+    detailed_stats_path = exp_dir / "detailed_run_stats.json"
+    detailed_stats_interface = DetailedJSONStatsInterface(detailed_stats_path)
+    stats = detailed_stats_interface[episode]
 
-    # Plot params.
-    ylims = [(0, 25), (75, 100)]
-    xticks = np.arange(5)
+    # Select a step where there's a goal state was achieved, and it was
+    # for pose estimation (object was already determined).
+    goal_states = get_goal_states(stats, achieved=True)
+    goal_states = [g for g in goal_states if g["info"]["is_pose_hypothesis"]]
+    gs = goal_states[0]
+    step = gs["info"]["matching_step_when_output_goal_set"]
 
-    # 1-LM
-    for ax_num, ax in enumerate([bottom_ax, top_ax]):
-        ax.bar(
-            xticks[0],
-            [performance.accuracy[0]],
-            color=TBP_COLORS["blue"],
-            width=0.8,
-            label="no voting",
-        )
-
-    # Multi-LM
-    for ax_num, ax in enumerate([bottom_ax, top_ax]):
-        ax.bar(
-            xticks[1:],
-            performance.accuracy[1:],
-            color=TBP_COLORS["purple"],
-            width=0.8,
-            label="voting",
-        )
-
-    for ax_num, ax in enumerate([bottom_ax, top_ax]):
-        ax.set_ylim(ylims[ax_num])
-
-    # Sets parameters for both x-axes (they're shared, so removing ticks for the
-    # top plot removes ticks for the bottom plot).
-    bottom_ax.set_xlabel("Num. LMs")
-    bottom_ax.set_xticks(xticks)
-    bottom_ax.set_xticklabels(["1", "2", "4", "8", "16"])
-    bottom_ax.legend(loc="lower right")
-
-    bottom_ax.set_ylabel("Accuracy (%)")
-    bottom_ax.set_yticks([0, 10, 20])
-    top_ax.spines.bottom.set_visible(False)
-    top_ax.set_yticks([80, 90, 100])
-
-    # Draw y-axis divider markers.
-    marker_kwargs = dict(
-        marker=[(-1, -0.5), (1, 0.5)],
-        markersize=8,
-        linestyle="none",
-        color="k",
-        mec="k",
-        mew=1,
-        clip_on=False,
+    # Get the pose MLHs.
+    mlh_graph_id = stats["LM_0"]["current_mlh"][step]["graph_id"]
+    top_mlh, second_mlh = get_top_two_mlhs_for_object(mlh_graph_id, stats, step)
+    fig, axes = plot_hypotheses_visualization(
+        stats,
+        step,
+        top_mlh,
+        second_mlh,
+        style=style,
     )
-    top_ax.plot([0], [0], transform=top_ax.transAxes, **marker_kwargs)
-    bottom_ax.plot([0], [1], transform=bottom_ax.transAxes, **marker_kwargs)
 
-    fig.savefig(out_dir / "accuracy.png")
-    fig.savefig(out_dir / "accuracy.svg")
+    # Add label, legends, etc.
+    view_angles = [(-37, 26, 0), (-7.50, 34.26, 0)]
+    for i, ax in enumerate(axes):
+        ax.set_proj_type("persp", focal_length=0.8)
+        axes3d_set_aspect_equal(ax)
+        axes3d_clean(ax)
+        ax.view_init(*view_angles[i])
+
+    # Add legend to the second plot.
+    sm = SensorModuleData(stats["SM_0"], style=style)
+    legend_handles = get_visualization_legend_handles(
+        ["sensor", "pose 1", "pose 2", "goal"], sm.style
+    )
+    axes[1].legend(
+        handles=legend_handles,
+        bbox_to_anchor=(0.1, 0.8),
+        framealpha=1,
+        fontsize=8,
+    )
+
+    fig.savefig(out_dir / "pose_hypotheses_visualization.png", bbox_inches="tight")
+    fig.savefig(out_dir / "pose_hypotheses_visualization.svg", bbox_inches="tight")
+    plt.show()
+
+    return fig, axes
+
+
+"""
+--------------------------------------------------------------------------------
+Panels D and F: Hypotheses Evidence
+--------------------------------------------------------------------------------
+"""
+
+
+def set_steps_above_threshold(
+    hypotheses: Iterable[Mapping],
+    x_percent_threshold: float = 20.0,
+) -> np.ndarray:
+    """Set the `steps_above_threshold` field for each hypothesis.
+
+    Helper function for `init_object_hypotheses` and `init_pose_hypotheses`.
+
+    Args:
+        hypotheses: Hypotheses to find steps above threshold for.
+        x_percent_threshold: Percentage threshold for evidence. Default is 20.
+
+    Returns:
+        np.ndarray: The step-wise evidence threshold.
+    """
+    # Find highest evidence at each step to compute the evidence threshold.
+    n_steps = hypotheses[0]["evidence"].shape[0]
+    highest_evidences = np.zeros(n_steps)
+    for step in range(n_steps):
+        highest_evidences[step] = np.max([h["evidence"][step] for h in hypotheses])
+    thresholds = highest_evidences * (1 - x_percent_threshold / 100)
+
+    # Find the last step where evidence was above threshold for each hypothesis.
+    for h in hypotheses:
+        where_above_threshold = np.where(h["evidence"] >= thresholds)[0]
+        if where_above_threshold.size == 0:
+            h["steps_above_threshold"] = 0
+        else:
+            h["steps_above_threshold"] = where_above_threshold[-1] + 1
+
+    return thresholds
+
+
+def init_object_hypotheses(stats: Mapping) -> np.ndarray:
+    """Initialize object hypotheses.
+
+    Args:
+        stats: Detailed stats for an episode.
+
+    Returns:
+        np.ndarray: An array of hypotheses, where each is a dictionary with the
+          following keys:
+            - `object_name`: The name of the object.
+            - `label`: The label for the hypothesis.
+            - `evidence`: The evidence for the hypothesis.
+    """
+    max_evidence_dicts: List[Mapping[str, float]] = stats["LM_0"]["evidences_max"]
+    hypotheses = []
+    for key in max_evidence_dicts[0].keys():
+        h = {
+            "object_name": key,
+            "label": key,
+            "evidence": np.array([dct[key] for dct in max_evidence_dicts]),
+        }
+        hypotheses.append(h)
+    hypotheses = np.array(hypotheses, dtype=object)
+    set_steps_above_threshold(hypotheses)
+    return hypotheses
+
+
+def init_pose_hypotheses(stats: Mapping, object_name: str) -> np.ndarray:
+    """Initialize pose hypotheses.
+
+    Args:
+        stats: Detailed stats for an episode.
+        object_name: The name of the object to initialize pose hypotheses for.
+
+    Returns:
+        np.ndarray: An array of hypotheses, where each is a dictionary with the
+          following keys:
+            - `rotation_matrix`: The rotation matrix for the hypothesis.
+            - `evidence`: The evidence for the hypothesis.
+            - `label`: The label for the hypothesis.
+    """
+    evidences = np.array([dct[object_name] for dct in stats["LM_0"]["evidences"]]).T
+    rotations = stats["LM_0"]["possible_rotations"][0][object_name]
+    hypotheses = []
+    for i in range(len(rotations)):
+        angles = R.from_matrix(rotations[i]).as_euler("xyz", degrees=True)
+        if np.any(angles < 0):
+            angles = (angles + 180) % 360
+        label = f"({int(angles[0])}, {int(angles[1])}, {int(angles[2])})"
+        h = {
+            "rotation_matrix": rotations[i],
+            "evidence": evidences[i],
+            "label": label,
+        }
+        hypotheses.append(h)
+    hypotheses = np.array(hypotheses, dtype=object)
+    set_steps_above_threshold(hypotheses)
+    return hypotheses
+
+
+def plot_goal_state_steps(ax: plt.Axes, goal_states: List[Mapping]):
+    for gs in goal_states:
+        step = gs["info"]["matching_step_when_output_goal_set"]
+        if gs["info"]["is_pose_hypothesis"]:
+            color = "gray"
+            ls = "--"
+        else:
+            color = "gray"
+            ls = "-"
+        ax.axvline(step, color=color, lw=1, linestyle=ls, alpha=1)
+
+
+def plot_evidence_for_one_hypothesis(
+    ax: plt.Axes,
+    h: Mapping,
+    skip_if_empty: bool = False,
+) -> None:
+    """Plot the evidence over time for a single hypothesis.
+
+    Helper function for `plot_hypotheses_evidence` and
+    """
+    last_step = h["steps_above_threshold"]
+    if last_step == 0 and skip_if_empty:
+        return
+    arr = h["evidence"][:last_step]
+    ax.plot(arr, **h["style"])
+
+
+def plot_hypotheses_evidence(
+    ax: plt.Axes,
+    hypotheses: Iterable[Mapping],
+    n_top_hypotheses: int = 4,
+    n_bottom_hypotheses: Optional[Number] = np.inf,
+    bottom_hypothesis_color: str = "gray",
+    bottom_hypothesis_alpha: float = 0.5,
+    legend_kwargs: Optional[Mapping] = None,
+) -> None:
+    # Split hypotheses into two groups.
+    top_hypotheses = hypotheses[:n_top_hypotheses]
+    if n_bottom_hypotheses is None:
+        bottom_hypotheses = []
+    elif n_bottom_hypotheses is np.inf:
+        bottom_hypotheses = hypotheses[n_top_hypotheses:]
+    else:
+        h_start = n_top_hypotheses
+        h_stop = n_top_hypotheses + n_bottom_hypotheses
+        bottom_hypotheses = hypotheses[h_start:h_stop]
+
+    # Add style to hypotheses.
+    for i, h in enumerate(top_hypotheses):
+        h["style"] = {
+            "color": HYPOTHESIS_COLORS[i],
+            "alpha": 1,
+            "label": h["label"],
+        }
+    for i, h in enumerate(bottom_hypotheses):
+        h["style"] = {
+            "color": bottom_hypothesis_color,
+            "alpha": bottom_hypothesis_alpha,
+            "label": None,
+        }
+        if i == 0:
+            h["style"]["label"] = "others"
+
+    # Plot in reverse order because zorder argument doesn't work.
+    for i, h in enumerate(bottom_hypotheses[::-1]):
+        can_skip = h["style"]["label"] is None
+        plot_evidence_for_one_hypothesis(ax, h, skip_if_empty=can_skip)
+    for h in top_hypotheses[::-1]:
+        plot_evidence_for_one_hypothesis(ax, h, skip_if_empty=False)
+
+    # Hack to get legend items in order since we had to plot in reverse order.
+    legend_order = [h["style"]["label"] for h in top_hypotheses]
+    if len(bottom_hypotheses) > 0:
+        legend_order.append("others")
+    cur_handles, cur_labels = ax.get_legend_handles_labels()
+    handles, labels = [], []
+    for lbl in legend_order:
+        try:
+            ind = cur_labels.index(lbl)
+            handles.append(cur_handles[ind])
+            labels.append(cur_labels[ind])
+        except ValueError:
+            pass
+    legend_kwargs = legend_kwargs or {}
+    ax.legend(handles, labels, **legend_kwargs)
+
+
+def finalize_evidence_axes(
+    ax: plt.Axes,
+    jump_step: Optional[int] = None,
+    **kw,
+):
+    # Set axes properties.
+    for key, val in kw.items():
+        getattr(ax, f"set_{key}")(val)
+
+    # Add a triangle indicating where the jump was. Has to be done last.
+    if jump_step is not None:
+        draw_triangle(
+            ax,
+            jump_step,
+            ax.get_ylim()[1],
+            theta=180,
+            radius=2,
+            align="bottom",
+            facecolor=TBP_COLORS["yellow"],
+            edgecolor="black",
+        )
+
+
+def plot_object_hypotheses_evidence():
+    """Plot evidence over time for object hypotheses.
+
+    Requires having run `fig5_hypothesis_driven_policy`.
+
+    Output is saved to `DMC_ANALYSIS_DIR/fig5/object_hypotheses`.
+    """
+    # Initialize ouput directory.
+    out_dir = OUT_DIR / "object_hypotheses"
+    out_dir.mkdir(parents=True, exist_ok=True)
+
+    # Initialize stats.
+    episode = 0
+    exp_dir = VISUALIZATION_RESULTS_DIR / "fig5_hypothesis_driven_policy"
+    detailed_stats_path = exp_dir / "detailed_run_stats.json"
+    detailed_stats = DetailedJSONStatsInterface(detailed_stats_path)
+    stats = detailed_stats[episode]
+
+    # Get all object hypotheses.
+    hypotheses = init_object_hypotheses(stats)
+    n_steps = hypotheses[0]["evidence"].shape[0]
+
+    # Collect match threshold info.
+    # Sort by evidence at step of hypothesis-driven jump.
+    goal_states = get_goal_states(stats)
+    obj_goal_states = [g for g in goal_states if not g["info"]["is_pose_hypothesis"]]
+    if len(obj_goal_states) == 0:
+        order_step = n_steps - 1
+    else:
+        order_step = obj_goal_states[0]["info"]["matching_step_when_output_goal_set"]
+    sort_by = [h["evidence"][order_step] for h in hypotheses]
+    hypotheses = hypotheses[np.argsort(sort_by)[::-1]]
+
+    fig, ax = plt.subplots(1, 1, figsize=(3.75, 2.5), dpi=300)
+    plot_goal_state_steps(ax, goal_states)
+    plot_hypotheses_evidence(
+        ax,
+        hypotheses,
+        4,
+        np.inf,
+        legend_kwargs={"title": "Hypothesis", "fontsize": 8, "handlelength": 0.75},
+    )
+    # Finalize axes.
+    finalize_evidence_axes(
+        ax,
+        jump_step=order_step,
+        xlim=(0, 15),
+        ylim=(0, 30),
+        xticks=[0, 5, 10, 15],
+        yticks=[0, 10, 20, 30],
+        xlabel="Step",
+        ylabel="Evidence",
+    )
+
+    fig.savefig(out_dir / "object_hypotheses_evidence.png", bbox_inches="tight")
+    fig.savefig(out_dir / "object_hypotheses_evidence.svg", bbox_inches="tight")
     plt.show()
 
 
-def save_performance_table() -> None:
-    """Save the performance metrics to a CSV file."""
-    out_dir = OUT_DIR / "performance"
-    out_dir.mkdir(exist_ok=True, parents=True)
-    experiments = [
-        "dist_agent_1lm_randrot_noise",
-        "dist_agent_2lm_randrot_noise",
-        "dist_agent_4lm_randrot_noise",
-        "dist_agent_8lm_randrot_noise",
-        "dist_agent_16lm_randrot_noise",
-    ]
-    performance = aggregate_multilm_performance_data(experiments)
-    performance = performance.drop(columns=["n_steps", "rotation_error"])
-    performance.to_csv(out_dir / "performance.csv")
+def plot_pose_hypotheses_evidence():
+    """Plot evidence over time for pose hypotheses.
+
+    Requires having run `fig5_hypothesis_driven_policy`.
+
+    Output is saved to `DMC_ANALYSIS_DIR/fig5/pose_hypotheses`.
+    """
+    # Initialize ouput directory.
+    out_dir = OUT_DIR / "pose_hypotheses"
+    out_dir.mkdir(parents=True, exist_ok=True)
+
+    # Initialize stats.
+    episode = 1
+    object_name = "mug"
+    exp_dir = VISUALIZATION_RESULTS_DIR / "fig5_hypothesis_driven_policy"
+    detailed_stats_path = exp_dir / "detailed_run_stats.json"
+    detailed_stats = DetailedJSONStatsInterface(detailed_stats_path)
+    stats = detailed_stats[episode]
+
+    # Get all object hypotheses.
+    hypotheses = init_pose_hypotheses(stats, object_name)
+    n_steps = hypotheses[0]["evidence"].shape[0]
+
+    # Collect match threshold info.
+    # Sort by evidence at step of hypothesis-driven jump.
+    goal_states = get_goal_states(stats)
+    pose_goal_states = [g for g in goal_states if g["info"]["is_pose_hypothesis"]]
+    if len(pose_goal_states) == 0:
+        order_step = n_steps - 1
+    else:
+        order_step = pose_goal_states[0]["info"]["matching_step_when_output_goal_set"]
+    sort_by = [h["evidence"][order_step] for h in hypotheses]
+    hypotheses = hypotheses[np.argsort(sort_by)[::-1]]
+
+    fig, ax = plt.subplots(1, 1, figsize=(3.75, 2.5), dpi=300)
+    plot_goal_state_steps(ax, goal_states)
+    plot_hypotheses_evidence(
+        ax,
+        hypotheses,
+        4,
+        np.inf,
+        legend_kwargs={"title": "Hypothesis", "fontsize": 8, "handlelength": 0.75},
+    )
+
+    # Finalize axes.
+    finalize_evidence_axes(
+        ax,
+        jump_step=order_step,
+        xlim=(0, 20),
+        xticks=[0, 5, 10, 15, 20],
+        ylim=(0, 30),
+        yticks=[0, 10, 20, 30],
+        xlabel="Step",
+        ylabel="Evidence",
+    )
+    fig.savefig(out_dir / "pose_hypotheses_evidence.png", bbox_inches="tight")
+    fig.savefig(out_dir / "pose_hypotheses_evidence.svg", bbox_inches="tight")
+    plt.show()
 
 
 if __name__ == "__main__":
-    plot_8lm_patches()
-    plot_steps()
-    plot_accuracy()
-    save_performance_table()
+    plot_curvature_guided_policy()
+    plot_performance()
+    plot_object_hypotheses_visualization()
+    plot_object_hypotheses_evidence()
+    plot_pose_hypotheses_visualization()
+    plot_pose_hypotheses_evidence()
